Remove Compatibility, Contributing and Development (#110)
* Remove Compatibility, Contributing and Development

Now that we have a wiki in my eyes we can remove those files. 
All links in Readme should now point to the wiki.

I'll still create this PR if anybody has a problem with deleting

* Delete COMPATIBILITY.md

See commit above

* Delete CONTRIBUTING.md

See commit above.

* Re-Add CONTRIBUTING.md

Re-added Contributing page, it now basically only is a link to the wiki page on contributing.